+++ src/pages/FavoriteRestaurants.js
@@ -1,0 +1,5 @@
+export default function FavoriteRestaurants(){
+    return (
+        <h1>Favorites</h1>
+    )
+}
