+++ src/pages/Home.js
@@ -1,0 +1,5 @@
+export default function FavoriteRestaurants(){
+    return (
+        <h1>Home</h1>
+    )
+}
